Make a separate file for add contents function
